requirements.txt
streamlit 
pandas 
numpy
plotly-express
